requirements.txt
fastapi
uvicorn
PyMuPDF
pymongo
python-multipart
pydantic
python-dotenv
pillow
easyocr
pdf2image
transformers
requests
pytest
httpx
pytest-asyncio